requirements.txt
discord.py[voice]
Flask
python-dotenv
webdavclient3
requests